+++ src/component.vue
@@ -50,6 +50,10 @@
           wrap: false
         })
       },
+      events: {
+        type: Array,
+        default: () => hooks
+      }
     },
     data() {
       return {
@@ -64,8 +68,8 @@
       /* istanbul ignore if */
       if (this.fp) return;
 
-      // Inject our method into hooks array
-      hooks.forEach((hook) => {
+      // Inject our method into events array
+      this.events.forEach((hook) => {
         this.config[hook] = arrayify(this.config[hook] || []).concat((...args) => {
           this.$emit(camelToKebab(hook), ...args)
         });
@@ -104,7 +108,8 @@
       config: {
         deep: true,
         handler(newConfig) {
-          // Workaround: Don't pass hooks to configs again
+          // Workaround: Don't pass hooks to configs again otherwise
+          // previously registered hooks will stop working
           hooks.forEach((hook) => {
             delete newConfig[hook];
           });